[Refactor] - #13 RegisterDeliveryCharge 분리
