+++ Napzakmarket-iOS/Napzakmarket-iOS/Presentation/Register/View/SellRegisterView.swift
@@ -12,25 +12,11 @@ struct SellRegisterView: View {
     @State private var description = ""
     @State private var productState = ""
     @State private var price = ""
-    
-    
-    
-    // 택배비 포함
-    @State private var deliveryChargeFree = true
-    
-    // 일반 택배
-    @State private var normalDelivery = false
-    @State private var normalDeliveryCharge = ""
-    
-    // 반값/알뜰 택배
-    @State private var halfDelivery = false
-    @State private var halfDeliveryCharge = ""
-    
-    
-
-    
-    // 최대 금액 100만원
-    private let maxPrice: Int = 1_000_000
+    @State private var deliveryChargeFree = true        // 배달비 여부
+    @State private var normalDelivery = false           // 일반 배달비 선택 여부
+    @State private var normalDeliveryCharge = ""        // 일반 배달비 금액
+    @State private var halfDelivery = false             // 알뜰,반값 배달비 선택 여부
+    @State private var halfDeliveryCharge = ""          // 알뜰,반값 배달비 금액
 
     var body: some View {
         NavigationStack {
@@ -60,157 +46,18 @@ struct SellRegisterView: View {
                         .padding(.vertical, 40)
                     
                     // MARK: - 가격
-                    
-                    VStack(alignment: .leading) {
-                        Text("가격")
-                            .padding(.bottom, 12)
-                        
-                        HStack{
-                            TextField("0", text: $price)
-                                .onChange(of: price) { oldValue, newValue in
-                                    price = stringToPrice(input: newValue)
-                                }
-                            
-                            Text("원")
-                                .foregroundColor(.gray)
-                        }
-                        .padding(.horizontal, 14)
-                        .padding(.vertical, 11)
-                        .overlay {
-                            RoundedRectangle(cornerRadius: 12)
-                                .stroke(.gray, lineWidth: 1)
-                        }
-                        
-                    }
-                    .padding(.bottom, 35)
-                    .padding(.horizontal, 20)
+                    RegisterSellPrice(price: $price)
 
                     
                     // MARK: - 배송비
+                    RegisterDeliveryCharge(
+                        deliveryChargeFree: $deliveryChargeFree,
+                        normalDelivery: $normalDelivery,
+                        normalDeliveryCharge: $normalDeliveryCharge,
+                        halfDelivery: $halfDelivery,
+                        halfDeliveryCharge: $halfDeliveryCharge
+                    )
                     
-                    VStack(alignment: .leading, spacing: 12) {
-                        Text("배송비")
-                        
-                        HStack{
-                            Button {
-                                deliveryChargeFree = true
-                            } label: {
-                                Text("포함")
-                                    .frame(maxWidth: .infinity, minHeight: 40)
-                                    .clipShape(.rect(cornerRadius: 12))
-                                    .background(deliveryChargeFree ? .white : .clear)
-                                    .clipShape(.rect(cornerRadius: 12))
-                                    .foregroundStyle(deliveryChargeFree ? .black : .gray)
-                            }
-                            
-                            Button {
-                                deliveryChargeFree = false
-                            } label: {
-                                Text("별도")
-                                    .frame(maxWidth: .infinity, minHeight: 40)
-                                    .background(deliveryChargeFree ? .clear : .white)
-                                    .clipShape(.rect(cornerRadius: 12))
-                                    .foregroundStyle(deliveryChargeFree ? .gray : .black)
-                            }
-                            
-                        }
-                        .padding(3)
-                        .background(.gray.opacity(0.2))
-                        .clipShape(.rect(cornerRadius: 12))
-                        
-                        if !deliveryChargeFree {
-                            
-                            // 일반 택배
-                            VStack {
-                                HStack {
-                                    Image(systemName: normalDelivery ? "checkmark.square.fill" : "square")
-                                        .foregroundColor(normalDelivery ? .purple : .gray)
-                                    
-                                    Text("일반 택배")
-                                    
-                                    Spacer()
-                                    
-                                    Image(systemName: normalDelivery ? "chevron.up" : "chevron.down")
-                                        .foregroundColor(.gray)
-                                }
-                                .onTapGesture {
-                                    normalDelivery.toggle()
-                                }
-                                
-                                if normalDelivery {
-                                    VStack {
-                                        HStack {
-                                            TextField("0", text: $normalDeliveryCharge)
-                                                .onChange(of: normalDeliveryCharge) { oldValue, newValue in
-                                                    normalDeliveryCharge = stringToPrice(input: newValue)
-                                                }
-                                            
-                                            Text("원")
-                                                .foregroundColor(.gray)
-                                        }
-                                        .padding(.horizontal, 15)
-                                        .padding(.vertical, 11)
-                                        .background(Color.gray.opacity(0.1))
-                                        .clipShape(.rect(cornerRadius: 10))
-                                    }
-                                }
-                                
-                            }
-                            .padding(.horizontal, 15)
-                            .padding(.vertical, 20)
-                            .overlay {
-                                RoundedRectangle(cornerRadius: 12)
-                                    .stroke(.gray, lineWidth: 1)
-                            }
-                            
-                            // 알뜰/반값 택배
-                            VStack {
-                                HStack {
-                                    Image(systemName: halfDelivery ? "checkmark.square.fill" : "square")
-                                        .foregroundColor(halfDelivery ? .purple : .gray)
-                                    
-                                    Text("알뜰/반값 택배")
-                                    
-                                    Spacer()
-                                    
-                                    Image(systemName: halfDelivery ? "chevron.up" : "chevron.down")
-                                        .foregroundColor(.gray)
-                                }
-                                .onTapGesture {
-                                    halfDelivery.toggle()
-                                }
-                                
-                                if halfDelivery {
-                                    VStack {
-                                        HStack {
-                                            TextField("0", text: $halfDeliveryCharge)
-                                                .onChange(of: halfDeliveryCharge) { oldValue, newValue in
-                                                    halfDeliveryCharge = stringToPrice(input: newValue)
-                                                }
-                                            
-                                            Text("원")
-                                                .foregroundColor(.gray)
-                                        }
-                                        .padding(.horizontal, 15)
-                                        .padding(.vertical, 11)
-                                        .background(Color.gray.opacity(0.1))
-                                        .clipShape(.rect(cornerRadius: 10))
-                                    }
-                                }
-                                
-                            }
-                            .padding(.horizontal, 15)
-                            .padding(.vertical, 20)
-                            .overlay {
-                                RoundedRectangle(cornerRadius: 12)
-                                    .stroke(.gray, lineWidth: 1)
-                            }
-                            
-                        }
-                        
-                    }
-                    .padding(.horizontal, 20)
-
                 }
                 .padding(.bottom, 67)
             }
@@ -232,23 +79,7 @@ struct SellRegisterView: View {
         
         
     }
-    
-    
-    // MARK: - 3자리마다 쉼표 추가 함수
-    private func stringToPrice(input: String) -> String {
-        // 숫자만 남기기
-        let filteredPrice = input.filter { $0.isNumber }
-        guard let price = Int(filteredPrice) else { return "" }
-        
-        // 최대 금액 제한
-        let limitedPrice = min(price, maxPrice)
-        
-        // 숫자를 3자리마다 쉼표로 구분
-        let formatter = NumberFormatter()
-        formatter.numberStyle = .decimal
-        return formatter.string(from: limitedPrice as NSNumber) ?? ""
-    }
-    
+
 }
 
 #Preview {
